Add more functions, fix build
